move modals to styleguide
