+++ package.json
@@ -32,6 +32,7 @@
     "@testing-library/react": "^10.4.3",
     "@testing-library/user-event": "^12.0.11",
     "@types/jest": "^26.0.10",
+    "@types/react-dom": "^17.0.8",
     "@types/systemjs": "^6.1.0",
     "autoprefixer": "^9.6.1",
     "babel-eslint": "^11.0.0-beta.2",
@@ -55,8 +56,8 @@
     "fork-ts-checker-webpack-plugin": "^6.2.6",
     "html-webpack-plugin": "^5.3.1",
     "husky": "^6.0.0",
-    "identity-obj-proxy": "^3.0.0",
     "i18next-parser": "^4.0.1",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^26.4.2",
     "jest-cli": "^26.4.2",
     "lerna": "^3.20.2",
