--- packages/framework/esm-styleguide-react/.babelrc
@@ -1,7 +1,0 @@
-{
-  "presets": [
-    "@babel/preset-env",
-    "@babel/preset-typescript",
-    "@babel/preset-react"
-  ]
-}
