--- packages/framework/esm-styleguide-react/README.md
@@ -1,2 +1,0 @@
-# openmrs-esm-styleguide-react
-
