--- packages/framework/esm-styleguide-react/package.json
@@ -1,65 +1,0 @@
-{
-  "name": "@openmrs/esm-styleguide",
-  "version": "3.1.10",
-  "license": "MPL-2.0",
-  "description": "The styleguide for OpenMRS SPA",
-  "browser": "dist/openmrs-esm-styleguide.js",
-  "main": "src/index.ts",
-  "source": true,
-  "scripts": {
-    "document": "../../../document.sh esm-styleguide",
-    "test": "jest --passWithNoTests",
-    "start:storybook": "start-storybook --port 7000",
-    "build:storybook": "build-storybook",
-    "build": "webpack --mode=production",
-    "analyze": "webpack --mode=production --env.analyze=true",
-    "lint": "eslint src"
-  },
-  "keywords": [
-    "openmrs",
-    "microfrontends",
-    "styleguide"
-  ],
-  "directories": {
-    "lib": "dist",
-    "src": "src"
-  },
-  "browserslist": [
-    "extends browserslist-config-openmrs"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/openmrs/openmrs-esm-core.git"
-  },
-  "bugs": {
-    "url": "https://github.com/openmrs/openmrs-esm-core/issues"
-  },
-  "homepage": "https://github.com/openmrs/openmrs-esm-core#readme",
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "@carbon/charts": "^0.41.57",
-    "carbon-components-react": "7.31.0",
-    "lodash-es": "^4.17.21"
-  },
-  "peerDependencies": {
-    "carbon-components": "10.x",
-    "carbon-icons": "7.x",
-    "react": "16.x",
-    "react-dom": "16.x",
-    "rxjs": "6.x"
-  },
-  "devDependencies": {
-    "@pickra/copy-code-block": "^1.1.7",
-    "@storybook/addon-a11y": "^5.2.1",
-    "@storybook/addon-knobs": "^5.2.1",
-    "@storybook/html": "^5.2.1",
-    "carbon-components": "10.31.0",
-    "carbon-icons": "7.0.7",
-    "highlight.js": "^9.15.10",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "rxjs": "^6.5.3"
-  }
-}
