--- packages/framework/esm-styleguide-react/postcss.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  plugins: [
-    require("autoprefixer"),
-    require("cssnano")({
-      preset: "default",
-    }),
-  ],
-};
