--- packages/framework/esm-styleguide-react/src/declarations.d.tsx
@@ -1,3 +1,0 @@
-declare module '*.css';
-declare module '*.scss';
-declare module '*.png';
