--- packages/framework/esm-styleguide-react/src/index.ts
@@ -1,1 +1,0 @@
-export { default as Modal } from "./Modal";
