--- packages/framework/esm-styleguide-react/tsconfig.json
@@ -1,24 +1,0 @@
-{
-  "compilerOptions": {
-    "esModuleInterop": true,
-    "module": "esnext",
-    "allowSyntheticDefaultImports": true,
-    "jsx": "react",
-    "strictNullChecks": true,
-    "moduleResolution": "node",
-    "declaration": true,
-    "declarationDir": "dist",
-    "emitDeclarationOnly": true,
-    "lib": [
-      "dom",
-      "es5",
-      "scripthost",
-      "es2015",
-      "es2015.promise",
-      "es2016.array.include",
-      "es2018",
-      "esnext"
-    ]
-  },
-  "include": ["src/**/*"]
-}
