+++ packages/framework/esm-styleguide/src/_all.scss
@@ -19,5 +19,6 @@
 @import "./notifications/notifications";
 @import "./toasts/toasts";
 @import "./components/forms";
+@import "./modals/modals";
 
 @import "./overrides";
