+++ packages/framework/esm-styleguide/src/declarations.d.ts
@@ -1,3 +1,4 @@
+declare module "*.scss";
 declare module "*.svg" {
   const content: string;
   export = content;
