+++ packages/framework/esm-styleguide/src/index.ts
@@ -5,6 +5,7 @@ export * from "./breakpoints";
 export * from "./spinner";
 export * from "./notifications";
 export * from "./toasts";
+export * from "./modals";
 
 setupLogo();
 setupIcons();
