+++ packages/framework/esm-styleguide/src/modals/_modals.scss
@@ -1,4 +1,4 @@
-.modal {
+.omrs-modal {
   position: fixed;
   width: 100vw;
   height: 100vh;
